stepSelectors in globalDefaults to allow for different ways of defining and finding impress steps ( using the 'step' class is very generic and likely to cause collisions with other libraries. preference is to use a specific data-impress-step attribute.
getImpressDataAttributes() to allow for mutationobserving on the attributes that are relevant to the impressjs code.

getPrevStep() can now skip the substeps ( fixes a bug with data-inherit="prev" )

delay getting the dataset from the step element until the value is needed. this will allow for recalculation if the actual data attributes get changed on the html element. (Useful for dynamic additions of steps )